move admin stuff to admin.py
